modeling players

+++ src/domain/entity/team/team.go
@@ -8,7 +8,7 @@ import (
 
 // Team is a team which is playing the cup
 type Team struct {
-	ID             int                `gorm:"primary_key;auto_increment" json:"id"`
+	ID             uint64             `gorm:"primary_key;auto_increment" json:"id"`
 	Name           string             `gorm:"not null" json:"name"`
 	Nickname       string             `gorm:"not null" json:"nickname"`
 	ImageShield    string             `gorm:"size:255;null;" json:"image_shield"`
